Complete datareaders and data struct

+++ truck_and_drone/include/datahandling/instance.h
@@ -4,6 +4,6 @@
 struct Instance {
     int n;
     int lim;
-    std::vector<std::vector<float>> truck_matrix;
-    std::vector<std::vector<float>> drone_matrix;
+    std::vector<std::vector<long long>> truck_matrix;
+    std::vector<std::vector<long long>> drone_matrix;
 };
